requirements.txt
discord.py~=1.7.2
PyMoe~=1.0.6
better-profanity~=0.7.0
aiohttp~=3.7.4.post0